Reworking the structure of the crate
Move all parsers into a sentences directory.
Move all commonly shared parsers into a utils.rs file.
Rework the RMC parser so it fits the way we use nom in the other parsers
more.
